fix: component naming mismatcheas
Fix component naming mismatches
